+++ docs/changes.rst
@@ -12,6 +12,8 @@ Weblate 5.16
 * Added Anthropic machinery integration, see :ref:`mt-anthropic`.
 * Encoding for :ref:`formats` can now be configured using :ref:`file_format_params` (e.g., ``csv_encoding``, ``properties_encoding``).
 * Added support for anonymous commit names via :setting:`PRIVATE_COMMIT_NAME_TEMPLATE`.
+* Added :setting:`WEBSITE_ALERTS_ENABLED` setting to allow disabling website
+  availability checks and alerts.
 
 .. rubric:: Improvements